Make datasource filtering consistent
Symphony passed different data to the `buildDSRetrievalSQL()` function depending on the filter query. Think of the following cases:

    key-equals: colour=blue,red
    key-equals: colour=blue,red + key-ranges: size=1..10

In the first case, Symphony will pass the following data for the `key-equals` filter:

```php
array(
	'key-equals: colour=blue',
	'red'
)
```

In the latter case, where the query contains a `+` operator (AND), the same `key-equals` statement will result in the following data:

```php
array(
	'key-equals: colour=blue,red'
)
```

The current implementation will return a working database query but it will create one join per colour value, when it's supposed to create only a single join. If you filter on a large number of entries, a list of only a few colours will result in a timeout leaving open connections that will quickly exceed the maximum connections of the database server, if your site is hosted on a shared host.

This change fixes this issue and will normalise the query building.